add CMouseClicker Class to simulate mouse click, when bird will crash pipe in about 0.3s the crash time is simple forecast by height / current speed, it's unprecise
CMouseClicker only clicks when mouse is in Game Canvas Area
bird height data type in CHeightTimeModel changes to percent float, for better compatible in various canvas scales

+++ FlappyBirdAI/util/CPerformanceCounter.cpp
@@ -12,14 +12,14 @@ void InitPreciseTickCount()
 }
 
 
-LONGLONG GetPreciseTickCount() {
+double GetPreciseTickCount() {
     LARGE_INTEGER llStart;
     QueryPerformanceCounter(&llStart);
-    return (LONGLONG)(llStart.QuadPart * QUADPART);
+    return (double)(llStart.QuadPart * QUADPART);
 }
 
 
-void LogPerformance(const char* name, LONGLONG tickcount)
+void LogPerformance(const char* name, double tickcount)
 {
-    DLOG(INFO) << "PerformaceCounter: " << name << ": " << (unsigned int)tickcount;
+    DLOG(INFO) << "PerformaceCounter: " << name << ": " << (double)tickcount;
 }
